Refine deadline refresh and centralize test helpers
